start dot_process

+++ ft_cal_digit_bonus.c
@@ -1,16 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_cal_digit.c                                     :+:      :+:    :+:   */
+/*   ft_cal_digit_bonus.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ttachi <ttachi@student.42tokyo.ja>         +#+  +:+       +#+        */
+/*   By: ttachi <ttachi@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/27 10:03:19 by ttachi            #+#    #+#             */
-/*   Updated: 2022/11/27 10:51:18 by ttachi           ###   ########.fr       */
+/*   Updated: 2022/11/27 16:23:34 by ttachi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-# include "ft_printf.h"
+#include "ft_printf.h"
 
 int	ft_cal_digit(long long num)
 {
@@ -27,7 +27,7 @@ int	ft_cal_digit(long long num)
 	return (digit);
 }
 
-int	ft_cal_hex_digit(unsigned int num, int digit)
+int	ft_cal_hex_digit(unsigned long num, int digit)
 {
 	if (num < 16)
 		return (++digit);
